demopython/frontend/requirements.txt
Flask==2.1.0
requests==2.23.0
pylint==2.4.0
gunicorn==20.0.4
pymongo==3.11.0
Flask-PyMongo==2.3.0
prometheus-flask-exporter
opentracing==2.4.0
jaeger-client==4.3.0
Flask-OpenTracing==1.1.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-exporter-jaeger